requirements.txt
fastapi
uvicorn
python-docx
pytesseract
PyMuPDF
sentence-transformers
transformers
torch
python-multipart
python-dotenv
google-generativeai
Pillow
